pyproject.toml
[tool.poetry]
name = "Strelix Identity"
version = "0.0.1"
description = "github.com/Strelix/Identity"
authors = ["TreyWW"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
bleach = "6.1.0"  # used for HTML sanitation
tblib = "^3.0.0"  # Exceptions and Tracebacks
django = "^5.0.7"
redis = { extras = ["hiredis"], version = "^5.0.1" }
djangorestframework = "^3.15.1"
drf-yasg = "^1.21.7"
gunicorn = "22.0.0"
social-auth-app-django = "5.4.1"
django-environ = "0.11.2"
django-extensions = "3.2.3"
python-dateutil = "2.9.0"
django-mathfilters = "1.0.0"
django-ratelimit = "4.1.0"
python-memcached = "1.62"
django-htmx = "1.17.3"
django-login-required-middleware = "0.9.0"
django-components = "0.37"
django-markdownify = "0.9.3"
shortuuid = "1.0.13"
model-bakery = "1.17.0"

setuptools = "^70.1.1"
djangorestframework-api-key = "^3.0.0"
[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
djlint = "^1.34.1"
coverage = "^7.4.4"
pre-commit = "^3.7.0"
bump-my-version = "^0.19.3"
mypy = "1.7.1"
django-stubs = { version = "4.2.7" }
django-stubs-ext = { version = "4.2.7" }
djangorestframework-stubs = { extras = ["compatible-mypy"], version = "^3.14.5" }
djangorestframework = "^3.14.0"
typos = "^1.20.3"
types-python-dateutil = "^2.9.0.20240906"
types-markdown = "^3.7.0.20240822"
types-pygments = "^2.18.0.20240506"
types-cffi = "^1.16.0.20240331"
types-protobuf = "^5.27.0.20240907"
types-psycopg2 = "^2.9.21.20240819"
types-pycurl = "^7.45.3.20240421"
types-six = "^1.16.21.20240513"
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.15"
mkdocs-git-committers-plugin-2 = "^2.3.0"

[tool.poetry.group.mysql]
optional = true

[tool.poetry.group.mysql.dependencies]
mysqlclient = "2.2.0"
mariadb = "1.1.8"

[tool.poetry.group.postgres]
optional = true

[tool.poetry.group.postgres.dependencies]
psycopg2 = "2.9.9"

[tool.djlint]
profile = "django"
max_line_length = 110

[tool.black]
line-length = 140

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.bumpversion]
current_version = "0.9.1"
commit = true
commit_args = "-s"
tag = true
tag_name = "v{new_version}"
sign_tags = true
allow_dirty = true
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)(\\.(?P<dev>post)\\d+\\.dev\\d+)?"
serialize = [
    "{major}.{minor}.{patch}.{dev}{$PR_NUMBER}.dev{distance_to_latest_tag}",
    "{major}.{minor}.{patch}"
]

message = "Version updated from {current_version} to {new_version}"

[[tool.bumpversion.files]]
filename = "backend/__init__.py"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[tool.mypy]
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main"
]
disable_error_code = ["annotation-unchecked"]
explicit_package_bases = true
incremental = false
exclude = [
    "venv",
    ".venv",
    "env",
    "ENV",
    "ENV_DIR",
    "env.bak",
    "venv.bak",
]

[tool.django-stubs]
django_settings_module = "settings.settings"

[[tool.mypy.overrides]]
module = [
    "django.core.cache.backends.redis",
    "django_ratelimit.core",
    "django_ratelimit.decorators",
    "django_ratelimit.exceptions",
    "django_ratelimit",
    "environ",
    "login_required",
    "emails",
    "step_functions",
    "django_components",
    "drf_yasg",
    "drf_yasg.utils",
    "drf_yasg.views",
]
ignore_missing_imports = true
